+++ src/pages/Lend/LoanChannelCategoryControl.vue
@@ -75,7 +75,7 @@
 					</template>
 
 					<helpme-choose-wrapper
-						v-if="enableHelpmeChoose"
+						v-if="showHelpMeChooseFeat"
 						:remaining-loans="helpmeChooseRemainingLoans"
 						:items-in-basket="itemsInBasket"
 						:is-visitor="isVisitor"
@@ -137,7 +137,7 @@
 					</template>
 
 					<helpme-choose-wrapper
-						v-if="enableHelpmeChoose"
+						v-if="showHelpMeChooseFeat"
 						:remaining-loans="helpmeChooseRemainingLoans"
 						:items-in-basket="itemsInBasket"
 						:is-visitor="isVisitor"
@@ -397,7 +397,7 @@ export default {
 			return (this.loanChannel?.loans?.values ?? []).filter(loan => loan !== null);
 		},
 		loans() {
-			if (this.enableHelpmeChoose) {
+			if (this.showHelpMeChooseFeat) {
 				return _filter(this.allLoans, (loan, index) => index < 6);
 			}
 			return this.allLoans;
@@ -407,7 +407,7 @@ export default {
 			return this.allLoans[0] ? [this.allLoans[0]] : [];
 		},
 		remainingLoans() {
-			if (this.enableHelpmeChoose) {
+			if (this.showHelpMeChooseFeat) {
 				return _filter(this.allLoans, (loan, index) => index > 0 && index < 5);
 			}
 			return _filter(this.allLoans, (loan, index) => index > 0);
@@ -461,6 +461,9 @@ export default {
 		hasOnePageOfLoans() {
 			return this.totalCount <= this.limit;
 		},
+		showHelpMeChooseFeat() {
+			return this.enableHelpmeChoose && this.allLoans.length > 8;
+		}
 	},
 	apollo: {
 		preFetch(config, client, args) {